server/requirements.txt
Flask
requests
PyGithub
python-dotenv
google-generativeai